+++ secretspec/src/provider/mod.rs
@@ -83,6 +83,8 @@ pub mod keyring;
 pub mod lastpass;
 pub mod onepassword;
 pub mod pass;
+#[cfg(feature = "vault")]
+pub mod vault;
 #[macro_use]
 pub mod macros;